loginserver: determine which channel IP to push - local or remote We have channel IP defined in gameserver(channel server) config and up until now we used to push this exact value to the channel list.
One of cons is, that if LAN IP is used there, then no remote connections can connect.

However, we do have private IP ranges and we can check if our client is connecting
from such network, and if it does - push local IP, otherwise use from the config one.

Here we assume that if client is connecting from private IP range, then all servers will
be on the same IP address as loginserver is. In other words, this will push loginserver's IP.